fix: Reduce scale of in‑scene objects and fix broken scene transitions.
- 씬들의 전체 오브젝트를 플레이어 크기에 맞게 축소시킴
- 씬들의 장면 전환 오류를 해결함

-> 플레이어에 인스턴스 스크립트, 쉐이더, 레이어 설정 스크립트만 넣어주면 작동!
