+++ Road_Of_MartialArts/Assets/Prototype/P_Enemy/Ghost.prefab
@@ -124,7 +124,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 7.1, y: 147.9, z: 227.5}
-  m_LocalScale: {x: 1.34413, y: 1.34413, z: 1.34413}
+  m_LocalScale: {x: 0.4461015, y: 0.4461015, z: 0.4461015}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 6165959744191260325}
