+++ Road_Of_MartialArts/Assets/Prototype/P_Enemy/Moodang.prefab
@@ -27,7 +27,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -43.8, y: 148.56, z: 231.92}
-  m_LocalScale: {x: 1.6071198, y: 1.6071198, z: 1.6071198}
+  m_LocalScale: {x: 0.6592913, y: 0.6592913, z: 0.6592913}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 857289432304066366}
