+++ Road_Of_MartialArts/Assets/Prototype/P_Enemy/Scarecrow.prefab
@@ -125,7 +125,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 219.2, y: 9.96, z: 0}
-  m_LocalScale: {x: 1.9215356, y: 1.9215357, z: 1.9215357}
+  m_LocalScale: {x: 0.50845754, y: 0.50845754, z: 0.50845754}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 8750794120128816378}
