+++ Road_Of_MartialArts/Assets/Prototype/P_Object/NPC/JungJo.prefab
@@ -28,7 +28,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -66.8, y: -12.98, z: 66.67}
-  m_LocalScale: {x: 1.1723603, y: 1.1723603, z: 1.1723603}
+  m_LocalScale: {x: 0.48570696, y: 0.48570696, z: 0.48570696}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 4774226378462448807}
