+++ Road_Of_MartialArts/Assets/Prototype/P_Object/NPC/NPC1.prefab
@@ -125,7 +125,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0.009136628, w: 0.9999583}
   m_LocalPosition: {x: -75, y: -12.57, z: 63.32}
-  m_LocalScale: {x: 1.2944797, y: 1.2944797, z: 1.2944797}
+  m_LocalScale: {x: 0.47540757, y: 0.47540757, z: 0.47540757}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 17001541158425730}
