+++ Road_Of_MartialArts/Assets/Prototype/P_Object/NPC/NPC2.prefab
@@ -28,7 +28,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -4.3, y: -14.55, z: 68.3}
-  m_LocalScale: {x: 1.6283994, y: 1.6283994, z: 1.6283994}
+  m_LocalScale: {x: 0.87272, y: 0.87272, z: 0.87272}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 303195074258680016}
