+++ Road_Of_MartialArts/Assets/Prototype/P_Object/NPC/NPC3.prefab
@@ -125,7 +125,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 12.550003, y: -12.66, z: 109.9}
-  m_LocalScale: {x: 2.8207006, y: 2.8207006, z: 2.8207006}
+  m_LocalScale: {x: 0.98755777, y: 0.98755777, z: 0.98755777}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 8755582583737099158}
