+++ Road_Of_MartialArts/Assets/Prototype/P_Object/NPC/NPC4.prefab
@@ -125,7 +125,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 3.8000011, y: -13.13, z: 112.02}
-  m_LocalScale: {x: 1.8731349, y: 1.8731349, z: 1.8731349}
+  m_LocalScale: {x: 0.7216792, y: 0.7216792, z: 0.7216792}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 7863700363344855323}
