+++ Road_Of_MartialArts/Assets/Prototype/P_Object/objects/Barrel_Dynamic.prefab
@@ -125,7 +125,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 19.2, y: -17.7, z: 75.77}
-  m_LocalScale: {x: 0.7220074, y: 0.7220074, z: 0.7220074}
+  m_LocalScale: {x: 0.3488238, y: 0.3488238, z: 0.3488238}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 778255422043191272}
