+++ Road_Of_MartialArts/Assets/Prototype/P_Object/objects/Haitai.prefab
@@ -124,7 +124,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -32.382965, y: -13.93, z: 69.66}
-  m_LocalScale: {x: 3.7537026, y: 3.7537026, z: 3.7537026}
+  m_LocalScale: {x: 1.5635356, y: 1.5635356, z: 1.5635356}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 3544600634709636507}
