+++ Road_Of_MartialArts/Assets/Prototype/P_Object/objects/Joseon_bg01.prefab
@@ -124,7 +124,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -98.5, y: 8.8, z: 94.7}
-  m_LocalScale: {x: 23.856491, y: 23.856491, z: 23.856491}
+  m_LocalScale: {x: 15.0669, y: 15.0669, z: 15.0669}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 3115371139387493144}
