+++ Road_Of_MartialArts/Assets/Prototype/P_Object/objects/signpost2.prefab
@@ -27,7 +27,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalScale: {x: 0.42035344, y: 0.42035344, z: 0.42035344}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1625507630443727570}
