+++ Road_Of_MartialArts/Assets/Prototype/P_Scene/P_Scripts/Transition/ExitPoint.cs
@@ -8,18 +8,11 @@ public class ExitPoint : MonoBehaviour
     [SerializeField]
     private string transferSceneName;
 
-    private Move_JS player;
-
-    private void Start()
-    {
-        player = Move_JS.instance;
-    }
-
     private void OnTriggerEnter(Collider other)
     {
         if (other.CompareTag("Player"))
         {
-            player.currentSceneName = SceneManager.GetActiveScene().name;
+            Player_instance_JS.instance.saveSceneName = SceneManager.GetActiveScene().name;
             SceneManager.LoadScene(transferSceneName);
         }
     }
