+++ Road_Of_MartialArts/Assets/Prototype/P_Scene/P_Scripts/Transition/NextPointZ.cs
@@ -7,24 +7,15 @@ public class NextPointZ : MonoBehaviour
 { 
     public Transform target;
 
-    private Move_JS player;
-    private CameraMove_JS cam;
-
-    private void Start()
-    {
-        player = Move_JS.instance;
-        cam = CameraMove_JS.instance;
-    }
-
     private void OnTriggerEnter(Collider other)
     {
         if (other.CompareTag("Player"))
         {
-            CharacterController cc = player.GetComponent<CharacterController>();
+            CharacterController cc = Player_instance_JS.instance.GetComponent<CharacterController>();
             cc.enabled = false; // 임시로 비활성화
             Vector3 newPos = new Vector3(other.transform.position.x, target.transform.position.y, target.transform.position.z);
-            player.transform.position = newPos;
-            cam.SnapToTarget();
+            Player_instance_JS.instance.transform.position = newPos;
+            CameraMove_JS.instance.SnapToTarget();
             cc.enabled = true;  // 다시 활성화
         }
     }
