+++ Road_Of_MartialArts/Assets/_Dev2/Field/Scripts/Player/Player_instance_JS.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ecf4bd416da83634a8f76efda16d7a59
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
